requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
pyyaml==6.0.1
pyarrow==16.0.0
kagglehub==0.2.9
